aifinepro/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
requests
pillow
pydantic
python-multipart
numpy
opencv-python
aiohttp
asyncio
google-genai